04-multimodal/live/pyproject.toml
[project]
name = "telegram-llm-bot"
version = "0.1.0"
description = "Simple Telegram bot with LLM integration"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.15.0",
    "openai>=1.54.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "ffmpeg-python>=0.2.0",
    "soundfile>=0.12.1",
    "vosk>=0.3.45",
]

